+++ attribute/benchmark_test.go
@@ -28,8 +28,8 @@ var (
 func benchmarkEmit(kv attribute.KeyValue) func(*testing.B) {
 	return func(b *testing.B) {
 		b.ReportAllocs()
-		for i := 0; i < b.N; i++ {
-			outStr = kv.Value.Emit()
+		for b.Loop() {
+			kv.Value.Emit()
 		}
 	}
 }
@@ -267,6 +267,25 @@ func BenchmarkStringSlice(b *testing.B) {
 		b.ReportAllocs()
 		for i := 0; i < b.N; i++ {
 			outStrSlice = kv.Value.AsStringSlice()
+		}
+	})
+	b.Run("Emit", benchmarkEmit(kv))
+}
+
+func BenchmarkEmpty(b *testing.B) {
+	k := "empty"
+	kv := attribute.Empty(k)
+
+	b.Run("Value", func(b *testing.B) {
+		b.ReportAllocs()
+		for b.Loop() {
+			attribute.EmptyValue()
+		}
+	})
+	b.Run("KeyValue", func(b *testing.B) {
+		b.ReportAllocs()
+		for b.Loop() {
+			attribute.Empty(k)
 		}
 	})
 	b.Run("Emit", benchmarkEmit(kv))