+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The next release will require at least [Go 1.25].
 ### Added
 
 - Support testing of [Go 1.26]. (#7902)
+- Add `Empty` and `EmptyValue` functions for new `EMPTY Type` in `go.opentelemetry.io/otel/attribute`.
+  Note that the zero value of `Value` is still invalid. (#7942)
 
 <!-- Released section -->
 <!-- Don't change this section unless doing release -->